ft_printf

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: kato <kato@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/30 20:51:28 by kato              #+#    #+#             */
-/*   Updated: 2025/05/04 18:07:14 by kato             ###   ########.fr       */
+/*   Updated: 2025/05/04 18:21:33 by kato             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ static int	handle_format(char c, va_list args)
 	else if (c == 'x' || c == 'X')
 		return (ft_putnbr_hex(va_arg(args, unsigned int), c));
 	else if (c == 'p')
-		return (ft_putptr(va_arg(args, void *)));
+		return (ft_putptr((unsigned long)va_arg(args, void *)));
 	else if (c == '%')
 		return (ft_putchar('%'));
 	return (0);
